+++ android/app/src/main/java/com/created/team201/data/mapper/ProfileMapper.kt
@@ -1,8 +1,8 @@
 package com.created.team201.data.mapper
 
-import com.created.domain.model.ProfileInformation
 import com.created.domain.model.Nickname
 import com.created.domain.model.Profile
+import com.created.domain.model.ProfileInformation
 import com.created.team201.data.remote.response.MyPageResponseDto
 
 fun MyPageResponseDto.toDomain(): Profile = Profile(
