--- android/app/src/main/java/com/created/team201/presentation/common/TodoClickListener.kt
@@ -1,10 +1,0 @@
-package com.created.team201.presentation.common
-
-import com.created.team201.presentation.home.model.TodoWithRoundIdUiModel
-
-interface TodoClickListener {
-
-    fun clickOnNecessaryTodoCheck(todo: TodoWithRoundIdUiModel, isDone: Boolean)
-
-    fun clickOnOptionalTodoCheck(todo: TodoWithRoundIdUiModel, isDone: Boolean)
-}
