--- android/app/src/main/java/com/created/team201/presentation/home/HomeClickListener.kt
@@ -1,13 +1,0 @@
-package com.created.team201.presentation.home
-
-import com.created.team201.presentation.common.TodoClickListener
-import com.created.team201.presentation.home.model.TodoWithRoundIdUiModel
-
-interface HomeClickListener : TodoClickListener {
-
-    override fun clickOnNecessaryTodoCheck(todo: TodoWithRoundIdUiModel, isDone: Boolean)
-
-    override fun clickOnOptionalTodoCheck(todo: TodoWithRoundIdUiModel, isDone: Boolean)
-
-    fun clickOnStudyCard(studyId: Long)
-}
