+++ android/app/src/main/java/com/created/team201/presentation/home/adapter/viewholder/DashboardViewHolder.kt
@@ -3,46 +3,23 @@ package com.created.team201.presentation.home.adapter.viewholder
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.created.team201.databinding.ItemHomeBinding
-import com.created.team201.presentation.home.HomeClickListener
-import com.created.team201.presentation.home.adapter.HomeGrassAdapter
-import com.created.team201.presentation.home.adapter.OptionalToDoAdapter
-import com.created.team201.presentation.home.model.StudyUiModel
+import com.created.team201.databinding.ItemHomeDashBoardBinding
+import com.created.team201.presentation.home.model.HomeViewState
+
 
 class DashboardViewHolder(
-    onClick: HomeClickListener,
-    private val binding: ItemHomeBinding,
+    private val binding: ItemHomeDashBoardBinding
 ) : RecyclerView.ViewHolder(binding.root) {
-    private val optionalTodoAdapter: OptionalToDoAdapter by lazy { OptionalToDoAdapter(onClick) }
-    private val homeGrassAdapter: HomeGrassAdapter by lazy { HomeGrassAdapter() }
 
-    init {
-        binding.onClick = onClick
-        attachOptionalTodoAdapter()
-        attachHomeGrassAdapter()
-    }
-
-    private fun attachOptionalTodoAdapter() {
-        binding.rvHomeOptionalTodoList.setHasFixedSize(true)
-        binding.rvHomeOptionalTodoList.itemAnimator = null
-        binding.rvHomeOptionalTodoList.adapter = optionalTodoAdapter
-    }
-
-    private fun attachHomeGrassAdapter() {
-        binding.rvHomeGrass.setHasFixedSize(true)
-        binding.rvHomeGrass.itemAnimator = null
-        binding.rvHomeGrass.adapter = homeGrassAdapter
-    }
-
-    fun bind(studyUiModel: StudyUiModel) {
-        binding.studyUiModel = studyUiModel
-
-        optionalTodoAdapter.submitList(studyUiModel.optionalTodos)
-        homeGrassAdapter.submitList(studyUiModel.grass)
+    fun bind(dashBoard: HomeViewState.DashBoard) {
+        binding.item = dashBoard
     }
 
     companion object {
-        fun getView(parent: ViewGroup, layoutInflater: LayoutInflater): ItemHomeBinding =
-            ItemHomeBinding.inflate(layoutInflater, parent, false)
+        fun from(parent: ViewGroup): DashboardViewHolder {
+            val binding = ItemHomeDashBoardBinding
+                .inflate(LayoutInflater.from(parent.context), parent, false)
+            return DashboardViewHolder(binding)
+        }
     }
 }
