+++ android/app/src/main/java/com/created/team201/presentation/home/adapter/viewholder/FeedViewHolder.kt
@@ -1,17 +1,17 @@
 package com.created.team201.presentation.home.adapter.viewholder
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.created.team201.databinding.ItemHomeFeedBinding
-import com.created.team201.presentation.home.adapter.HomeViewState.Feed
+import com.created.team201.presentation.home.model.HomeViewState
+
 
 class FeedViewHolder(
     private val binding: ItemHomeFeedBinding
 ) : RecyclerView.ViewHolder(binding.root) {
 
-    fun bind(feed: Feed) {
+    fun bind(feed: HomeViewState.Feed) {
         binding.item = feed
     }
