--- android/app/src/main/java/com/created/team201/presentation/home/model/TodoWithRoundIdUiModel.kt
@@ -1,10 +1,0 @@
-package com.created.team201.presentation.home.model
-
-data class TodoWithRoundIdUiModel(
-    val todoId: Long,
-    val content: String?,
-    val isDone: Boolean,
-    val roundId: Int,
-) {
-    fun isContentEmpty(): Boolean = !content.isNullOrEmpty()
-}
