+++ android/app/src/main/java/com/created/team201/presentation/studyManagement/model/NecessaryTodoUiModel.kt
@@ -1,7 +1,5 @@
 package com.created.team201.presentation.studyManagement.model
 
-import com.created.team201.presentation.home.model.TodoUiModel
-
 data class NecessaryTodoUiModel(
     val todo: TodoUiModel,
     val isInitialized: Boolean,
