+++ android/app/src/main/res/layout/activity_main.xml
@@ -29,7 +29,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fcv_main"
-            android:name="com.created.team201.presentation.home.HomeFragment2"
+            android:name="com.created.team201.presentation.home.HomeFragment"
             android:layout_width="0dp"
             android:layout_height="0dp"
             app:layout_constraintBottom_toTopOf="@id/bnv_main"
