--- android/app/src/main/res/layout/item_home_optional_todo.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="onClick"
-            type="com.created.team201.presentation.home.HomeClickListener" />
-
-        <variable
-            name="todoUiModel"
-            type="com.created.team201.presentation.home.model.TodoWithRoundIdUiModel" />
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/tv_home_optional_todo"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp"
-            android:text="@{todoUiModel.content}"
-            android:textAppearance="@style/text_r14"
-            android:textColor="@color/white"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="홈탭 작업 마무리하기" />
-
-        <ImageView
-            android:id="@+id/iv_home_optional_check_btn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="12dp"
-            android:onClick="@{() -> onClick.clickOnOptionalTodoCheck(todoUiModel,todoUiModel.isDone)}"
-            android:src="@drawable/ic_selector_home_check_btn"
-            app:isSelected="@{todoUiModel.done}"
-            app:isVisible="@{todoUiModel.contentEmpty}"
-            app:layout_constraintBottom_toBottomOf="@id/tv_home_optional_todo"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/tv_home_optional_todo" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
